==> analysis/18_kipoi_setup.ipynb <==
# Configurations
import warnings;warnings.filterwarnings("ignore")
from tensorflow.python.util import deprecation; deprecation._PRINT_DEPRECATION_WARNINGS = False

#Modules
import os
import sys
import numpy as np
import pandas as pd
from tqdm import tqdm
import plotnine
from plotnine import *
import seaborn as sns
import pickle
from glob import glob
from itertools import product, compress
from pybedtools import BedTool
from keras import backend as K

from datetime import datetime

from bpnet.utils import read_json, create_tf_session
from bpnet.dataspecs import DataSpec
from bpnet.datasets import StrandedProfile
from bpnet.extractors import StrandedBigWigExtractor
from bpnet.BPNet import BPNetSeqModel
from bpnet.metrics import eval_profile

#Setup
os.chdir('/l/Zeitlinger/ZeitlingerLab/Manuscripts/Zelda_and_Nucleosomes/Analysis/analysis/')
# create_tf_session('0')
%matplotlib inline

#Variables
prefix = 'ZDTBCG'
figure_filepath = f'figures/2_binding_bpnet_training'

model = BPNetSeqModel.from_mdir('bpnet/models/optimized_model/fold1')

model.seqmodel.model.summary()
